requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.0
qdrant-client==1.6.9
redis==5.0.1
google-generativeai==0.3.2
PyPDF2==3.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiosqlite==0.19.0
httpx==0.25.2
email-validator==2.1.0
requests==2.31.0
numpy==1.24.3
python-dateutil==2.8.2